default value passed to edit event form/ button component added w/styling

+++ src/routes/events_pages/EventDetailPage.js
@@ -7,26 +7,36 @@ import PageHeading from "../../components/ui/PageHeading";
 import EventDetails from "../../components/events_elements/EventDetails";
 import MainContentWrapper from "../../components/wrapper/MainContentWrapper";
 // fetched data
-import fetchEvents from "../../utility/fetch-events-data";
 import fetchEvent from "../../utility/fetch-event-data";
 
 const EventDetailPage = () => {
-  const {event} = useRouteLoaderData('events-details');
+  const { event } = useRouteLoaderData("events-details");
 
   const submit = useSubmit();
 
   const deleteEventHandler = () => {
-    submit(null, {method: 'delete'})
+    const proceed = window.confirm(
+      "Are you sure you want to delete this event?"
+    );
+    if (proceed) {
+      submit(null, { method: "delete" });
+    }
   };
 
   return (
     <MainContentWrapper>
-    <Suspense>
-      <PageHeading header={"Event Details"} />
-      <Await resolve={event}>
-        {(event) => <EventDetails deleteEvent={deleteEventHandler} name={event.name} date={event.date} />}
-      </Await>
-    </Suspense>
+      <Suspense>
+        <PageHeading header={"Event Details"} />
+        <Await resolve={event}>
+          {(event) => (
+            <EventDetails
+              deleteEvent={deleteEventHandler}
+              name={event.name}
+              date={event.date}
+            />
+          )}
+        </Await>
+      </Suspense>
     </MainContentWrapper>
   );
 };
@@ -36,26 +46,28 @@ export default EventDetailPage;
 export const loader = async ({ request, params }) => {
   const id = params.eventId;
 
-  console.log(id)
+  console.log(id);
 
   return defer({
-    event: fetchEvent(id),
+    event: await fetchEvent(id),
   });
-
 };
 
-export const action = async ({request, params}) => {
-  const id = params.eventId
+export const action = async ({ request, params }) => {
+  const id = params.eventId;
 
-  console.log(request)
+  console.log(request);
 
-  const response = await fetch(`https://react-http-6cb96-default-rtdb.europe-west1.firebasedatabase.app/events/${id}.json`, {
-    method: request.method
-  });
+  const response = await fetch(
+    `https://react-http-6cb96-default-rtdb.europe-west1.firebasedatabase.app/events/${id}.json`,
+    {
+      method: request.method,
+    }
+  );
 
   if (!response.ok) {
-    throw new Error("Something went wrong!")
-  } 
+    throw new Error("Something went wrong!");
+  }
 
-  return redirect('/events')
+  return redirect("/events");
 };
